+++ src/components/InputFile/Item/Item.tsx
@@ -9,20 +9,34 @@ type Props = {
   className?: string;
   img?: string;
   onRemove?: () => void;
-  loaded: number;
   total: number;
+  loaded: number;
+  waitingForUpload?: boolean;
 };
 
 export default class Item extends Component<Props> {
   render() {
-    const { className, img, loaded, total, onRemove, children } = this.props;
+    const {
+      className,
+      img,
+      total,
+      loaded,
+      waitingForUpload,
+      onRemove,
+      children,
+    } = this.props;
     const style = {} as CSSProperties;
     const isComplete = loaded === total;
+    const classes = cn(
+      S.root,
+      waitingForUpload && S.waitingForUpload,
+      className
+    );
 
     if (img) style.backgroundImage = `url(${img})`;
 
     return (
-      <div className={cn(S.root, className)} style={style}>
+      <div className={classes} style={style}>
         {children}
 
         <Button
@@ -37,7 +51,7 @@ export default class Item extends Component<Props> {
 
         <div
           className={cn(S.progress, isComplete && S.complete)}
-          style={{ left: `${(loaded / total) * 100}%` }}
+          style={{ left: `${(loaded / total) * 100 - 100}%` }}
         />
 
         <div className={S.overlay} />