+++ src/docs/examples/InputFile/index.tsx
@@ -1,6 +1,36 @@
+import { Container, Router, Link } from 'uilib';
+
 import { Code } from 'docs/components';
 
-import example from '!!raw-loader!./Example';
 import { upload } from './helpers';
 
-export default () => <Code code={example} scope={{ upload }} />;
+import UploadOnChange from '!!raw-loader!./UploadOnChange';
+import UploadOnDemand from '!!raw-loader!./UploadOnDemand';
+
+const rootPath = '/inputFile';
+
+export default () => (
+  <Container vertical fullWidth>
+    <div>
+      <Link href={`${rootPath}/onchange`}>Upload on change</Link>
+      <Link href={`${rootPath}/ondemand`}>Upload on demand</Link>
+    </div>
+
+    <Router rootPath={rootPath}>
+      <Code
+        exact
+        path="/onchange"
+        code={UploadOnChange}
+        scope={{ upload }}
+        key="onchange"
+      />
+      <Code
+        exact
+        path="/ondemand"
+        code={UploadOnDemand}
+        scope={{ upload }}
+        key="ondemand"
+      />
+    </Router>
+  </Container>
+);