+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.2.4](https://github.com/foreverido/uilib/compare/v1.2.3...v1.2.4) (2022-04-01)
+
+
+### Bug Fixes
+
+* Gallery - start at first, recenter() ([94b3f31](https://github.com/foreverido/uilib/commit/94b3f31482880d08377250c16c0952ff127d50db))
+
+### [1.2.3](https://github.com/foreverido/uilib/compare/v1.2.2...v1.2.3) (2022-03-31)
+
+
+### Bug Fixes
+
+* Router - getWeightestRoute() ([62f7e1f](https://github.com/foreverido/uilib/commit/62f7e1f829dc61c8826cfa5e4e57976a27759702))
+
+### [1.2.2](https://github.com/foreverido/uilib/compare/v1.2.1...v1.2.2) (2022-03-31)
+
+### [1.2.1](https://github.com/foreverido/uilib/compare/v1.2.0...v1.2.1) (2022-03-30)
+
+## [1.2.0](https://github.com/foreverido/uilib/compare/v1.1.1...v1.2.0) (2022-03-30)
+
+
+### Features
+
+* InputFile - upload onSubmit ([ad83d88](https://github.com/foreverido/uilib/commit/ad83d88fbf31233aa91877a3aa7fff156b93b1cb))
+* Lazy - allow define custom progressElem ([9ab779c](https://github.com/foreverido/uilib/commit/9ab779cf0649dfd3bff82cad6ec9a077de29e7b7))
+
 ### [1.1.1](https://github.com/foreverido/uilib/compare/v1.1.0...v1.1.1) (2022-03-29)